organize gantt component
