+++ src/components/Gantt/Gantt.js
@@ -1,19 +1,16 @@
 import React, { useEffect, useRef } from 'react';
 import { gantt } from 'dhtmlx-gantt';
-import ReactMarkdown from 'react-markdown';
-import ReactDOMServer from 'react-dom/server';
-
 import 'dhtmlx-gantt/codebase/dhtmlxgantt.css';
 import { getIssuesFromAPI } from '../../functions/Common/IssueAPI.js';
-import {
-  calculateDuration,
-  calculateDueDate,
-} from '../../functions/Common/CommonHelper.js';
+import { attachEvent } from './GanttAttachEvent.js';
+import { setGanttTemplates } from './GanttTemplates.js';
+import { setGanttConfig } from './GanttConfig.js';
 
 const Gantt = (props) => {
   const containerRef = useRef(null);
   useEffect(() => {
     setGanttConfig(gantt);
+    setGanttTemplates(gantt);
     attachEvent(gantt, props);
     gantt.init(containerRef.current);
     gantt.ext.zoom.setLevel(props.zoom);
@@ -58,185 +55,4 @@ const Gantt = (props) => {
   );
 };
 
-const zoom_level = {
-  levels: [
-    {
-      name: 'Days',
-      scale_height: 60,
-      min_column_width: 70,
-      scales: [{ unit: 'day', step: 1, format: '%n/%d' }],
-    },
-    {
-      name: 'Weeks',
-      scale_height: 60,
-      min_column_width: 70,
-      scales: [
-        { unit: 'month', step: 1, format: '%Y年 %n月' },
-        { unit: 'week', step: 1, format: '%n/%d~' },
-      ],
-    },
-  ],
-};
-
-const shortenDate = (date) => {
-  const m = ('00' + (date.getMonth() + 1)).slice(-2);
-  const d = ('00' + date.getDate()).slice(-2);
-  const shorten_date = m + '/' + d;
-  return shorten_date;
-};
-
-const setGanttConfig = (gantt) => {
-  gantt.config.xml_date = '%Y/%m/%d %H:%i';
-  gantt.config.order_branch = true;
-  gantt.config.order_branch_free = true;
-
-  gantt.config.keep_grid_width = true;
-  gantt.config.grid_resize = true;
-  gantt.config.open_tree_initially = true;
-  gantt.config.fit_tasks = true;
-
-  gantt.config.sort = true;
-  // gantt.config.show_task_cells = false;
-  // gantt.config.branch_loading = true;
-  // gantt.config.smart_scales = true;
-
-  gantt.config.columns = [
-    { name: 'wbs', label: 'WBS', width: 40, template: gantt.getWBSCode },
-    { name: 'id', label: 'No.', align: 'left', tree: true, width: '*' },
-    {
-      name: 'start_date',
-      label: 'Start ',
-      align: 'center',
-      width: '100',
-      template: (obj) => {
-        return shortenDate(obj.start_date);
-      },
-    },
-    {
-      name: 'due_date',
-      label: 'due ',
-      align: 'center',
-      width: '100',
-      template: (obj) => {
-        return shortenDate(obj.due_date);
-      },
-    },
-    { name: 'assignee', label: 'Assignee', align: 'center', width: '150' },
-    { name: 'add', label: '' },
-  ];
-
-  gantt.templates.timeline_cell_class = function (item, date) {
-    if (date < new Date()) {
-      return 'past_days';
-    }
-    if (date.getDay() === 0 || date.getDay() === 6) {
-      return 'weekend';
-    }
-  };
-
-  gantt.templates.rightside_text = function (start, end, task) {
-    return task.text;
-  };
-
-  gantt.templates.task_text = function (start, end, task) {
-    return (
-      "<span style='text-align:left;'>" +
-      Math.round(task.progress * 100) +
-      '% </span>'
-    );
-  };
-
-  gantt.templates.task_class = function (start, end, task) {
-    console.log(
-      calculateDueDate(start, calculateDuration(start, end) * task.progress)
-    );
-    if (
-      new Date(
-        calculateDueDate(
-          start,
-          (calculateDuration(start, end) + 1) * task.progress
-        )
-      ) < new Date()
-    ) {
-      return 'behind';
-    }
-    // return '';
-  };
-
-  gantt.plugins({
-    quick_info: true,
-    drag_timeline: true,
-  });
-  gantt.ext.zoom.init(zoom_level);
-};
-
-const attachEvent = (gantt, props) => {
-  gantt.attachEvent('onTaskDblClick', (gantt_task_id, e) => {
-    props.openIssueAtBrowser(gantt_task_id);
-  });
-  gantt.attachEvent('onTaskCreated', (gantt_task_id, e) => {
-    props.openNewIssueAtBrowser(gantt_task_id);
-  });
-  gantt.attachEvent('onAfterTaskUpdate', (id, gantt_task) => {
-    props.updateIssueByAPI(gantt_task, gantt);
-  });
-  gantt.attachEvent('onAfterTaskMove', (id, parent) => {
-    let gantt_task = gantt.getTask(id);
-    if ('parent' in gantt_task) {
-      if (gantt_task.parent !== 0) {
-        gantt_task.parent = parent;
-        props.updateIssueByAPI(gantt_task, gantt);
-      }
-    }
-  });
-  gantt.attachEvent('onQuickInfo', (id) => {
-    let gantt_task = gantt.getTask(id);
-    gantt.locale.labels.detail_button = 'DETAIL';
-    gantt.$click.buttons.detail_button = (gantt_task_id) => {
-      props.openIssueAtBrowser(gantt_task_id);
-      return true;
-    };
-
-    gantt.ext.quickInfo.setContent({
-      header: {
-        title: '',
-        date: ReactDOMServer.renderToStaticMarkup().toString(),
-      },
-      content: ReactDOMServer.renderToStaticMarkup(
-        <div>
-          <h3>{gantt_task.text}</h3>
-          <ReactMarkdown>{gantt_task.description}</ReactMarkdown>
-        </div>
-      ).toString(),
-      buttons: ['detail_button'],
-    });
-  });
-  
-  gantt.attachEvent('onTaskDrag', function (id, mode, task, original) {
-    var state = gantt.getState();
-    var minDate = state.min_date,
-      maxDate = state.max_date;
-
-    var scaleStep =
-      gantt.date.add(new Date(), state.scale_step, state.scale_unit) -
-      new Date();
-
-    var showDate,
-      repaint = false;
-    if (mode == 'resize' || mode == 'move') {
-      if (Math.abs(task.start_date - minDate) < scaleStep) {
-        showDate = task.start_date;
-        repaint = true;
-      } else if (Math.abs(task.end_date - maxDate) < scaleStep) {
-        showDate = task.end_date;
-        repaint = true;
-      }
-
-      if (repaint) {
-        gantt.render();
-        gantt.showDate(showDate);
-      }
-    }
-  });
-};
 export default Gantt;
